requirements.txt
selenium
webdriver-manager
pandas
openpyxl
tabula-py
requests
alive-progress
loguru